crewai_system/requirements.txt
# CrewAI核心依赖
crewai>=0.22.0
crewai-tools>=0.4.0

# 保持与原系统兼容的依赖
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-core>=0.3.0
pandas>=2.1.0
numpy>=1.24.0
python-dotenv>=1.0.0
matplotlib>=3.9.0
yfinance>=0.2.0
akshare>=1.11.0
requests>=2.31.0
beautifulsoup4>=4.12.0
openai>=1.12.0
google-generativeai>=0.3.0
backoff>=2.2.0

# FastAPI后端依赖
fastapi>=0.115.0
uvicorn>=0.34.0

# 开发和测试依赖
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.7.0
isort>=5.12.0
flake8>=6.1.0